+++ templates/esm/exceptions/UserNotFound.js
@@ -1,6 +1,0 @@
-export default class extends Error {
-  constructor() {
-    super('User not found');
-    this.name = 'UserNotFound';
-  }
-}