+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.7] - 2024/11/21
+### Changed
+- **Refactor:** Renamed the directory for storing custom error classes in the template from `exceptions` to `errors`.
+- Removed `errors/UserNotFound` from the template and added a more general-purpose `errors/NotFoundError` class.
+
 ## [2.0.6] - 2024/11/18
 ### Changed
 - Refactored variable names.
@@ -58,3 +63,4 @@ All notable changes to this project will be documented in this file.
 
 [2.0.5]: https://github.com/takuya-motoshima/express-sweet/compare/v2.0.4...v2.0.5
 [2.0.6]: https://github.com/takuya-motoshima/express-sweet/compare/v2.0.5...v2.0.6
+[2.0.7]: https://github.com/takuya-motoshima/express-sweet/compare/v2.0.6...v2.0.7